apps/ai/requirements-cpu.txt
# apps/ai/requirements.txt
# Python Dependencies for AI Service

# ============================================================================
# Core Framework
# ============================================================================
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# ============================================================================
# gRPC
# ============================================================================
grpcio
grpcio-tools
protobuf

# ============================================================================
# Machine Learning - Object Detection
# ============================================================================
# Machine Learning - Object Detection (CPU ONLY)
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.5.1+cpu
torchvision==0.20.1+cpu
facenet-pytorch
psycopg2-binary
ultralytics=8.3.15 #jupyter 

# torch
# torchvision
# facenet-pytorch
# psycopg2-binary
# ultralytics        # YOLOv11
#onnx==1.15.0
#onnxruntime-gpu==1.16.3     # Use onnxruntime for CPU-only

# ============================================================================
# Image Processing
# ============================================================================
opencv-python
#opencv-contrib-python==4.8.1.78
#Pillow==10.1.0
numpy

# ============================================================================
# Logging & Monitoring
# ============================================================================
structlog
#python-json-logger
#prometheus-client

# ============================================================================
# Utilities
# ============================================================================
#python-multipart==0.0.6
#python-dotenv==1.0.0

# ============================================================================
# Development & Testing (optional - see requirements-dev.txt)
# ============================================================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# httpx==0.25.2
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1